requirements.txt
bentoml>=1.4.0
scikit-learn>=1.8.0
pandas>=2.0.0
pyjwt>=2.10.0
python-dotenv>=1.0.0
pytest>=9.0.0
requests>=2.32.0

